requirements.txt
transformers
torch
huggingface-hub
neo4j
diffusers
Pillow
requests
accelerate
llama-cpp-python
outlines
python-dotenv